+++ packages/moonstone/Scrollable/useChild.js
@@ -26,7 +26,6 @@ const useChildAdapter = () => {
 	return [childAdapter, setChildAdapter];
 };
 
-export default useChildAdapter;
 export {
 	useChildAdapter
 };
