+++ packages/moonstone/Scrollable/useSpotlight.js
@@ -1,6 +1,37 @@
-import {useContext, useEffect} from 'react';
+import Spotlight from '@enact/spotlight';
+import utilDOM from '@enact/ui/Scrollable/utilDOM';
+import {useEffect} from 'react';
 
 import {SharedState} from '../internal/SharedStateDecorator/SharedStateDecorator';
+
+import scrollbarCss from './Scrollbar.module.less';
+
+const navigableFilter = (elem) => {
+	if (
+		!Spotlight.getPointerMode() &&
+		// ignore containers passed as their id
+		typeof elem !== 'string' &&
+		utilDOM.containsDangerously(elem.classList, scrollbarCss.scrollButton)
+	) {
+		return false;
+	}
+};
+
+const useSpotlightConfig = (props) => {
+	const {'data-spotlight-id': spotlightId, focusableScrollbar} = props;
+
+	// Hooks
+
+	useEffect(() => {
+		function configureSpotlight () {
+			Spotlight.set(spotlightId, {
+				navigableFilter: focusableScrollbar ? null : navigableFilter
+			});
+		}
+
+		configureSpotlight();
+	}, [focusableScrollbar, spotlightId]);
+};
 
 const useSpotlightRestore = (props, instances) => {
 	const {id} = props;
@@ -29,7 +60,7 @@ const useSpotlightRestore = (props, instances) => {
 	}, [context, id, uiScrollableAdapter]);
 };
 
-export default useSpotlightRestore;
 export {
+	useSpotlightConfig,
 	useSpotlightRestore
 };
